java/src/main/java/com/yourtion/leetcode/daily/m08/d12/Solution.java
package com.yourtion.leetcode.daily.m08.d12;


import com.yourtion.leetcode.utils.graph.Node;

import java.util.HashMap;

/**
 * 133. 克隆图
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/clone-graph/
 */
public class Solution {
    private final HashMap<Node, Node> visited = new HashMap<>();

    public Node cloneGraph(Node node) {
        if (node == null) {
            return null;
        }

        // 如果该节点已经被访问过了，则直接从哈希表中取出对应的克隆节点返回
        if (visited.containsKey(node)) {
            return visited.get(node);
        }

        // 克隆节点，注意到为了深拷贝我们不会克隆它的邻居的列表
        Node cloneNode = new Node(node.val);
        // 哈希表存储
        visited.put(node, cloneNode);

        // 遍历该节点的邻居并更新克隆节点的邻居列表
        for (Node neighbor : node.neighbors) {
            cloneNode.neighbors.add(cloneGraph(neighbor));
        }
        return cloneNode;
    }
}
